(refactor) confirm overlay
