+++ src/components/common/ConfirmOverlay/index.tsx
@@ -1,0 +1,75 @@
+import React from 'react'
+import styled from 'styled-components'
+
+const Wrapper = styled.div<{ direction?: string }>`
+  align-items: center;
+  background-color: rgba(255, 255, 255, 0.8);
+  bottom: 0;
+  display: flex;
+  flex-direction: ${(props) => props.direction};
+  justify-content: center;
+  left: 0;
+  position: absolute;
+  right: 0;
+  top: 0;
+  z-index: 123;
+`
+
+Wrapper.defaultProps = {
+  direction: 'column',
+}
+
+const ConfirmText = styled.p`
+  color: ${(props) => props.theme.colors.textColor};
+  font-size: 16px;
+  font-weight: 600;
+  line-height: 1.4;
+  margin: 0 0 8px;
+`
+
+const ConfirmControls = styled.div`
+  column-gap: 20px;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  margin: 0 auto;
+`
+
+const ConfirmControl = styled.span`
+  color: ${(props) => props.theme.colors.textColor};
+  cursor: pointer;
+  font-size: 15px;
+  font-weight: 700;
+  line-height: 1.4;
+  text-transform: uppercase;
+
+  &:hover {
+    text-decoration: underline;
+  }
+`
+
+const Yes = styled(ConfirmControl)`
+  color: ${(props) => props.theme.colors.error};
+`
+
+const No = styled(ConfirmControl)``
+
+interface Props {
+  confirm: () => void
+  deny: () => void
+  direction?: string
+  text: string
+}
+
+export const ConfirmOverlay: React.FC<Props> = (props) => {
+  const { confirm, deny, direction, text, ...restProps } = props
+
+  return (
+    <Wrapper direction={direction} {...restProps}>
+      <ConfirmText>{text}</ConfirmText>
+      <ConfirmControls>
+        <Yes onClick={confirm}>Yes</Yes>
+        <No onClick={deny}>No</No>
+      </ConfirmControls>
+    </Wrapper>
+  )
+}
